requirements.txt
pandas==1.0.4
python-dotenv==0.19.1
selenium==3.141.0
xlrd==1.2.0
XlsxWriter==3.0.1